+++ packages/common/src/types.ts
@@ -1153,6 +1153,7 @@ export enum AccountTransactionType {
     SimpleTransferWithMemo = 22,
     EncryptedTransferWithMemo = 23,
     TransferWithScheduleWithMemo = 24,
+    ConfigureBaker = 25,
     ConfigureDelegation = 26,
 }